% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{analyze_mlr_model}
\alias{analyze_mlr_model}
\title{MLR}
\usage{
analyze_mlr_model(df, dep_var, ind_vars)
}
\arguments{
\item{df}{A data frame.}

\item{dep_var}{A string of the dependent variable.}

\item{ind_vars}{A vector of strings of the independent variables names.}
}
\value{
A list containing the following components:
  \itemize{
    \item{reduced}{
      \itemize{
        \item{model_summary}{Summary of the reduced model.}
        \item{AIC}{Akaike Information Criterion for the reduced model.}
        \item{cooks_distance}{Cooks distance for the reduced model.}
        \item{summary_reduced_anova}{ANOVA summary for the reduced model.}
      }
    }
    \item{full}{
      \itemize{
        \item{model_summary}{Summary of the full model.}
        \item{AIC}{Akaike Information Criterion for the full model.}
        \item{cooks_distance}{Cooks distance for the full model.}
        \item{summary_full_anova}{ANOVA summary for the full model.}
      }
    }


  }
}
\description{
MLR
}
\examples{
data <- data.frame(Y = rnorm(50, mean = 50, sd = 10),
X1 = rnorm(50, mean = 5, sd = 2),
X2 = sample(c("A", "B", "C"), 50, replace = TRUE))
results <- analyze_mlr_model(data, "Y", c("X1", "X2"))

}
