% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boiler.R
\docType{data}
\name{boiler}
\alias{boiler}
\title{Data for Boiler Drum Study}
\format{
## `boiler`
A data frame with 36 rows and 6 columns:
\describe{
  \item{Man-Hours}{hour}
  \item{ Boiler Capacity}{thousand lbs/hr}
  \item{Design Pressure}{psi}
  ...
}
}
\source{
Textbook Example 11.11
}
\usage{
boiler
}
\description{
In a production facility, an accurate estimate of man-hours needed to complete
a task is crucial to management. A manufacturer of boiler drums wants to use
regression to predict the number of manhours needed to erect the drums in future projects.
}
\keyword{datasets}
