% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bub.R
\docType{data}
\name{bub}
\alias{bub}
\title{BUBBLE2 data set}
\format{
## `bub`
A data frame with 18 rows and 5 columns:
\describe{
  \item{MassFlux}{kilograms per meters squared per second}
  \item{HeatFlux}{megawatts per meters squared}
  \item{Diameter}{millimeters}
  \item{Density}{(liters per meters squared}
  ...
}
}
\source{
Textbook
}
\usage{
bub
}
\description{
A subset of data of bubble behavior in subcooled flow boiling
Report ...
}
\keyword{datasets}
