% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2Boots.R
\name{myboot}
\alias{myboot}
\title{Bootstrapping estimates}
\usage{
myboot(df, model, iter, alpha)
}
\arguments{
\item{df}{A data frame containing the data-set.}

\item{model}{A character string specifying the model ("model1" or "model2").}

\item{iter}{An integer specifying the number of bootstrap iterations.}

\item{alpha}{The significance level for confidence intervals.}
}
\value{
A list containing:
  \item{betas}{A matrix of bootstrapped estimates for model coefficients.}
  \item{point_estimates}{A vector of point estimates for model coefficients.}
  \item{confidence_intervals}{A data frame with lower and upper bounds of confidence intervals for model coefficients.}
}
\description{
Bootstrapping estimates
}
\examples{

set.seed(123)  # For reproducibility
sample_data <- data.frame(
  Diameter = rnorm(100, mean = 50, sd = 10),
  Density = rnorm(100, mean = 10, sd = 2),
  MassFlux = runif(100, min = 1, max = 5),
  HeatFlux = runif(100, min = 0, max = 10)
)
out_model1 <- myboot(df = sample_data, model = "model1", iter = 2000, alpha = 0.05)
str(out_model1)

out_model2 <- myboot(df = sample_data, model = "model2", iter = 2000, alpha = 0.05)
str(out_model2)


}
