% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3_MCMC_pack.R
\name{performAnalyses}
\alias{performAnalyses}
\title{Bayesian analysis}
\usage{
performAnalyses(data, formula, priorList = list(), interval = 0.95)
}
\arguments{
\item{data}{A data frame containing the variables for the regression analysis.}

\item{formula}{An object of class `formula` specifying the regression model.}

\item{priorList}{A list of prior settings for the Bayesian regression analysis.}

\item{interval}{The confidence level for the interval estimates in the classical analysis.}
}
\value{
A list containing:
  \item{classical_point_estimates}{Point estimates from the classical regression analysis.}
  \item{classical_interval_estimates}{Interval estimates from the classical regression analysis.}
  \item{bayesian_summary}{Summary of the Bayesian regression analysis.}
  \item{comparison}{Comparison of classical and Bayesian estimates.}
  \item{geweke_diag}{Geweke diagnostic for the Bayesian analysis.}
  \item{heidel_diag}{Heidelberger-Welch diagnostic for the Bayesian analysis.}
}
\description{
This function provides an approach for conducting both Bayesian and classical
linear regression analyses. It provides detailed output summaries, point and interval
estimates, comparative statistics, and diagnostic plots.
}
\examples{
df <- data.frame(temperature = c(16.8, 15.0, 16.5, 17.7, 20.6, 22.6, 23.3, 18.2, 18.6),
  catch_ratio = c(0.66, 0.30, 0.46, 0.44, 0.67, 0.99, 0.75, 0.24, 0.51))
a <- df$catch_ratio
b <- df$temperature
results <- performAnalyses(df, a ~ b + I(b^2) + I(b^3))
results2 <- performAnalyses(df, a ~ b + I(b^2))
}
