% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4_ShinyAPP.R
\name{runMLRShinyApp}
\alias{runMLRShinyApp}
\title{Shiny Interactive MLR Analysis}
\usage{
runMLRShinyApp(df)
}
\arguments{
\item{df}{A data frame containing the variables to be used in the MLR analysis.}
}
\value{
A Shiny app for interactive MLR analysis is launched.
}
\description{
Function launches an interactive Shiny app for performing MLR analysis on a provided data set.
Users can select response and predictor variables, choose the model type (Linear or Quadratic),
and optionally include interaction terms. The app dynamically generates regression models based
on user input, displaying both the regression plot and a summary of the model.
}
\examples{
\dontrun{df <- data.frame(temperature = c(16.8, 15.0, 16.5, 17.7, 20.6, 22.6, 23.3, 18.2, 18.6),
  catch_ratio = c(0.66, 0.30, 0.46, 0.44, 0.67, 0.99, 0.75, 0.24, 0.51))
runMLRShinyApp() }
\dontrun{
df <- data.frame(
  temperature = rnorm(100, mean = 20, sd = 5),
  humidity = rnorm(100, mean = 50, sd = 10),
  energy_usage = rnorm(100, mean = 200, sd = 20))}
}
